Rename `bases.py` to `aggregator_bases.py`
