+++ src/torchjd/aggregation/aligned_mtl.py
@@ -30,7 +30,7 @@ from torch import Tensor
 
 from ._utils.gramian import compute_gramian
 from ._utils.pref_vector import pref_vector_to_str_suffix, pref_vector_to_weighting
-from .bases import _WeightedAggregator
+from .aggregator_bases import _WeightedAggregator
 from .mean import _MeanWeighting
 from .weighting_bases import _Weighting
