+++ src/torchjd/aggregation/cagrad.py
@@ -9,7 +9,7 @@ from torch import Tensor
 
 from ._utils.gramian import compute_gramian, normalize
 from ._utils.non_differentiable import raise_non_differentiable_error
-from .bases import _WeightedAggregator
+from .aggregator_bases import _WeightedAggregator
 from .weighting_bases import _Weighting
