+++ src/torchjd/aggregation/constant.py
@@ -1,7 +1,7 @@
 from torch import Tensor
 
 from ._utils.str import vector_to_str
-from .bases import _WeightedAggregator
+from .aggregator_bases import _WeightedAggregator
 from .weighting_bases import _Weighting
