+++ src/torchjd/aggregation/nash_mtl.py
@@ -34,7 +34,7 @@ from cvxpy import Expression
 from torch import Tensor
 
 from ._utils.non_differentiable import raise_non_differentiable_error
-from .bases import _WeightedAggregator
+from .aggregator_bases import _WeightedAggregator
 from .weighting_bases import _Weighting
