+++ src/torchjd/aggregation/pcgrad.py
@@ -3,7 +3,7 @@ from torch import Tensor
 
 from ._utils.gramian import compute_gramian
 from ._utils.non_differentiable import raise_non_differentiable_error
-from .bases import _WeightedAggregator
+from .aggregator_bases import _WeightedAggregator
 from .weighting_bases import _Weighting
