+++ src/torchjd/aggregation/random.py
@@ -2,7 +2,7 @@ import torch
 from torch import Tensor
 from torch.nn import functional as F
 
-from .bases import _WeightedAggregator
+from .aggregator_bases import _WeightedAggregator
 from .weighting_bases import _Weighting
